// apps/self-signup/src/lib/clients/eventsAnalytics.ts
export type EventName =
  | 'Signup Initiated'
  | 'Create Account Submitted'
  | 'Questionnaire Initiated'
  | 'Questionnaire Proceeded'
  | 'Free Trial Initiated'
  | 'Email Id Entered'
  | 'Job Role Selected'
  | 'Industry Selected'
  | 'Company Size Selected'
  | 'Use Case Selected'
  | 'Company Name Entered'
  | 'Email Verification Retry'
  | 'Contact Sales selected'
  | 'Extend Free Trial Selected';

export type Category =
  | 'Signup Form'
  | 'Questionnaire'
  | 'Email Verification'
  | 'Expired Free Trial';